OpenScripts Update 2
[Python\Running Software Detection\Windows\WMIC]:

Fixed a bug where executables with spaces in their name are not detected as running when running. The bug is caused by the string split() method, which treats spaces as list separators. This has been fixed by replacing split() with splitlines() and additional condition statements to remove empty strings and additional whitespaces from existing strings.

Minor adjustments to the comments.

[Python\Running Software Detection\Windows\PowerShell]:
First stable release.